+++ prediction_market_agent_tooling/benchmark/utils.py
@@ -1,19 +1,9 @@
 import json
 import typing as t
-from datetime import datetime
 
 from pydantic import BaseModel
 
-from prediction_market_agent_tooling.markets.agent_market import (
-    AgentMarket,
-    FilterBy,
-    SortBy,
-)
 from prediction_market_agent_tooling.markets.data_models import Resolution
-from prediction_market_agent_tooling.markets.markets import (
-    MARKET_TYPE_TO_AGENT_MARKET,
-    MarketType,
-)
 
 
 class OutcomePrediction(BaseModel):
@@ -71,25 +61,6 @@ class PredictionsCache(BaseModel):
     def load(path: str) -> "PredictionsCache":
         with open(path, "r") as f:
             return PredictionsCache.model_validate(json.load(f))
-
-
-def get_binary_markets(
-    limit: int,
-    market_type: MarketType,
-    filter_by: FilterBy = FilterBy.OPEN,
-    sort_by: SortBy = SortBy.NONE,
-    excluded_questions: set[str] | None = None,
-    created_after: datetime | None = None,
-) -> t.List[AgentMarket]:
-    agent_market_class = MARKET_TYPE_TO_AGENT_MARKET[market_type]
-    markets = agent_market_class.get_binary_markets(
-        limit=limit,
-        sort_by=sort_by,
-        filter_by=filter_by,
-        created_after=created_after,
-        excluded_questions=excluded_questions,
-    )
-    return markets
 
 
 def get_llm_api_call_cost(
